+++ src/mol-model-formats/structure/property/bonds/index-pair.ts
@@ -2,6 +2,7 @@
  * Copyright (c) 2019-2022 Mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author Alexander Rose <alexander.rose@weirdbyte.de>
+ * @author David Sehnal <david.sehnal@gmail.com>
  */
 
 import { CustomPropertyDescriptor } from '../../../../mol-model/custom-property';
@@ -42,7 +43,7 @@ export namespace IndexPairBonds {
         name: 'index_pair_bonds',
     };
 
-    export const Provider = FormatPropertyProvider.create<IndexPairBonds>(Descriptor);
+    export const Provider = FormatPropertyProvider.create<IndexPairBonds>(Descriptor, { asDynamic: true });
 
     export type Data = {
         pairs: {